Add support for adding hosts to DNS
If a `domain_name` is configured in `vars`, then we should also add these hosts to DNS. Before adding, we will remove any stale records.
The hostname in `kafka_data` is used to generate self signed certificates. If we're using a FQDN name to connect to kafka, then the certificate needs to have this FQDN as well.